rename handleErrors.js
